BackEnd/requirements.txt
# FastAPI & Web Server
fastapi==0.125.0
uvicorn==0.38.0

# WebSocket & Socket.IO
python-socketio==5.13.0
python-engineio==4.12.2
simple-websocket==1.1.0
wsproto==1.2.0
h11==0.16.0
httpx[socks]==0.28.1

# CORS & Request Forms
python-multipart==0.0.21

# ORM & DB
sqlmodel==0.0.27
SQLAlchemy==2.0.41
pydantic==2.11.7
pydantic-settings==2.7.0
pydantic-core==2.33.2
typing-extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# Auth / JWT
PyJWT==2.10.1

# Timezone
pytz==2025.2

# Misc
click==8.1.8
colorama==0.4.6
bidict==0.23.1
python-dotenv==1.0.0

handyLLM
